requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
django-otp>=1.2
qrcode[pil]>=7.4
python-decouple>=3.8
django-axes>=6.0
Pillow>=10.0
